map: Extract offline tile logic into useOfflineTiles composable
Move shared offline tile download dialogs, controls, and progress
tracking from Map.vue and MissionPlanningView.vue into a reusable
composable, eliminating ~80 lines of duplication per file.
